server/requirements.txt
flask==2.0.1
flask-cors==3.0.10
flask-limiter==3.3.0
python-dotenv==0.19.0
requests==2.26.0
gunicorn==20.1.0 